Add dependencies and set --task as required arg
Wasn't sure if I should go with requirements.txt
or pyproject.toml for this. We don't need a lot of
the functionality of pyproject.toml, since we're not
building this as a shared library, etc... But if there
ever /is/ something from pyproject.toml that we need,
I didn't want to have to maintain two sources of truth
for what our dependencies were. But I'm still not 100%
clear on the difference between the two, so if anyone
has a better reason for one vs the other, let's hear it.

Setting --task as a required command line argument
cleans up the error message when you accidentally forget it.
Instead of "Could not find file .../None.json" you get
a message telling you that --task is required.