Built Overcooked AI dataset processor, and added overcooked_ai as a git submodule to check it out. The dataset processor supports both the .csv and .pickle files of the (cleaned!) overcooked datasets. It then goes through and transforms the  game State to an image, and recreates the dataset. Additionally for the pickle files, it recreates the pickle file using  our version of pandas for ease of processing in the rest of the repository.
File changes:
Updated the `OvercookedAIProcessor` class in `centralized_processor.py`. This processor processor converts symbolic
game states from the dataset into base64-encoded visual observations (images) using the `overcooked_ai_py` library.
A core challenge is that the Overcooked AI dataset contains pickle files generated with an older, incompatible version
of pandas.  As such we've added a fallback mechanism that creates a temporary virtual environment with the older pandas
and transforms the pickle file into a CSV for easier processing. This is a temporary solution as eventually pandas 1.5
will fall fully out of pip. There is a utility file added in src/utils/unpickle_helper.py, that handles this conversion
Additionally there is a script at src/utils/visualize_overcooked_data.py that can quickly render the csv file images
so that we can do a spot check of the visualized environments.

Finally there is a new test file in tst/test_overcooked_module.py, which has a basic test of the processor just for
future proofing and system analysis.